created field checks handler
